fix: improve invoice handling for the electronic invoice
This commit offers a more detailed pricing calculation, including Net and Gross pricing for the
electronic invoice export. It adds essential support for VAT calculation for invoice articles. It
also includes the quantity into the gross price calculation which was previously based only on the
net price. This could help in maintaining proper financial records. We also replaced the implicit
variable declaration with explicit variable assignment for better code readability.